pyproject.toml
[project]
name = "garmin-reports"
version = "0.1.0"
description = "Simple api to extract garmin data informations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=6.0.0",
    "click>=8.3.1",
    "fastapi>=0.124.4",
    "fastmcp>=2.14.1",
    "influxdb>=5.3.2",
    "loguru>=0.7.3",
    "polars>=1.35.2",
    "pydantic-ai-slim>=1.39.0",
    "python-dateutil>=2.9.0.post0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pydantic-ai-slim[mcp,openai]>=1.39.0",
    "ruff>=0.14.9",
]
